backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.13.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apscheduler==3.10.4
asyncpg==0.29.0
attrs==25.3.0
bcrypt==4.1.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
coverage==7.13.1
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==8.0.4
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.0
filetype==1.2.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.16
langchain-core==0.3.81
langchain-google-genai==2.1.5
langchain-ollama==0.2.2
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.5.0
mypy_extensions==1.1.0
narwhals==1.44.0
numpy==2.3.1
ollama==0.4.4
openai==2.14.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymilvus==2.6.6
pypdf==4.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.1.0
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.0.9
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
streamlit==1.46.0
structlog==24.1.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
